Added a new project file
This project is based on test data from a marketplace. The marketplace recently appeared on the market and sells new products from Brazil that are just starting to go on sale. The problem: the marketplace's revenue has been stagnant for several months. It is necessary to calculate monthly retention and understand which metrics the marketplace should focus on.